@@ -16,7 +16,7 @@
     "scripts": {
         "link": "sf link ./ -f",
         "build": "node ./scripts/build.mjs",
-        "test": "node ./scripts/test.cjs",
+        "test": "node ./scripts/test.mjs",
         "patch": "npm run test && npm run build && sf publish patch -r"
     },
     "files": [
@@ -27,13 +27,13 @@
         "type": "git",
         "url": "git+https://github.com/cenfun/console-grid.git"
     },
-    "dependencies": {},
     "devDependencies": {
-        "eastasianwidth": "^0.2.0",
-        "eight-colors": "^1.3.0",
-        "eslint": "^8.57.0",
-        "eslint-config-plus": "^1.0.6",
-        "eslint-plugin-html": "^8.0.0",
+        "chalk": "^5.4.1",
+        "eastasianwidth": "^0.3.0",
+        "eight-colors": "^1.3.1",
+        "eslint": "^9.17.0",
+        "eslint-config-plus": "^2.0.2",
+        "eslint-plugin-html": "^8.1.2",
         "js-beautify": "^1.15.1",
         "papaparse": "^5.4.1"
     }
